Agregue Controller,Service,Repository y Model de Inventario

+++ src/main/java/cl/perfulandia/productos/model/Categoria.java
@@ -10,8 +10,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
-
 @Entity
 @Table(name = "categoria")
 @Data
@@ -21,7 +19,7 @@ public class Categoria {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false, length = 100, unique = true)
     private String nombre;
