+++ src/main/java/cl/perfulandia/productos/model/Marca.java
@@ -1,4 +1,3 @@
-
 package cl.perfulandia.productos.model;
 
 import jakarta.persistence.Column;
@@ -20,7 +19,7 @@ public class Marca {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false, length = 100, unique = true)
     private String nombre;
